Classes: task, user Task implementation

+++ team-taskmanagement-core/src/main/java/com/team/taskmanagement/dao/CustomerDao.java
@@ -8,7 +8,7 @@ public interface CustomerDao {
 
     public void addCustomer (Customer customer);
 
-    public Customer findById (int id);
+    public Customer findById (long id);
 
     public List<Customer> myCustomerList();
